Wayland port
This change adds Wayland support only to Qt 5 builds (it's not possible
with Qt 4).

Users pass -DUSE_WAYLAND=ON to cmake in order to enable Wayland support.
Wayland and X11 support are mutually exclusive at this point.

systemd >= 209 is required because it manages devices for us.

Changes to daemon:

 - New WaylandDisplayServer class with Wayland specific code

Changes to greeter:

 - KeyboardBackend moved to its separate .cpp/.h files
 - Same happened to XcbKeyboardBackend
 - KeyboardModelPrivate moved to a separate header
 - Layout moved to separate .cpp/.h files as KeyboardLayout
 - New WaylandKeyboardBackend class with Wayland specific code
 - Added WaylandIntegration class to use wl_fullscreen_shell

Wayland backends are used only if USE_WAYLAND is defined, otherwise their
X counterpart is built and used.

The following sddm.conf entries are ignored with Wayland support:

 - ServerPath
 - SessionCommand
 - DisplayCommand

The sddm.conf configuration file has different values if Wayland support
is enabled for:

 - SessionsDir

With Wayland we have compositors that take roles the role of display server,
window manager and compositor in the X11 world.

There are several Wayland compositors, each with different command line
arguments.  For this reason the only compositor supported at this moment
is Weston with fullscreen shell, a plugin that is available since the 1.5
release.

Weston's fullscreen shell requires a window to be presented on each
output, size mismatch between window and output sizes is resolved
stretching the window to accomodate output size (other options are
available, though).
sddm currently creates only one giant window for all the screens,
to avoid stretching this window I decided to make it as big as
the primary output resulting in other outputs showing a black
screen.

In the future we might implement sddm's own compositor, but for an
initial Wayland support it's better use Weston.
A sddm compositor might be able to show sddm giant window without
problems.