Add support for display setup script
It is done the same way as KDM does it:
With autologin, it will be ran right after the display server starts.
Without autologin, it will be ran right after the greeter has connected to the daemon. This is needed, otherwise it'll just
show a black screen on an optimus laptop.