Make the greeter go through logind
Use QAuth for the greeter session so that it will be registered
as a logind session with class "greeter".

With this, logind can pass drm devices to the greeter,
resolving all permission denied errors.

This work will also be needed by the Wayland port.

From now on we don't support running the greeter as root,
however the old code is preserved for test mode.