Introduce runtime directory and use it for Xauthority
Runtime directory is meant for temporary files so use it to store the
Xauthority file.  We leave state directory regardless because it might
be useful in the future.

It's not configurable by the user to avoid messing around with delicate
configuration options and gives us a predictable location.

However in test mode we usually run as an unprivileged user who doesn't
have access to system-wide directories so we try to find a suitable path.

First we look for the current user runtime directory, if it's not
available we try with the temporary directory.