sddm-helper: exit with code 0 otherwise it looks like an internal error
When sddm-helper exits with a code other than 0, SDDM::Auth emits the
finished(false) signal which is interpreted as an internal error condition
by the daemon.

However with authentication errors we don't need to exit with an
error code, we just send an error through the socket. It will be
interpreted correctly by the daemon which will display an error in the
greeter.

Issue: #221