Don't restart greeter and display server if sddm-helper exited successfully
sddm-helper now finishes with exit code 0 even when an authentication failure
happend because it's a managed situation.

When an authentication error is raised we should really do something when
the error() signal is emitted.

Invoke stop() only when an internal error happens to avoid restarting the
greeter and display server for authentication errors, which is annoying
to say the least.

Issue: #221