Distinguish between sddm-helper exit codes
Restarting only in case of error like we did with the previous changes
leads to the greeter not being restarted when a user session ends.

The daemon need a way to know when it's appropriate to restart greeter and
display server according to what the helper says.

Making sddm-helper always return 0 in case of authentication error leads us
to confusion just like it was before, so it's better if it returns the
appropriate exit code and let the daemon decide what to do.