Fix running SDDM under a Plasma 5.0 desktop.
When running SDDM under a full Plasma 5.0 desktop, the integration plugin tries
to access the arguments to figure out a configuration file name.  However, as
argc in SDDM::GreeterApp was passed as a temporary into a QGuiApplication, which
takes argc as a reference, the reference to argc is no longer valid and the
application segfaults.  Fix by having GreeterApp take argc as a reference.