Don't override pam environment variables
When using pam_env users are allowed to provide environment variables
to SDDM.

This is often used to customize PATH like Debian does.
However we overwrite those environment variables with our own.

What happens on Debian is that we overwrite PATH with a less
complete version.

Make sure pam variables are not overridden.

Closes #256

[ChangeLog][Behavioral Change] Don't override pam environment
variables.