Load first theme in alphabetical order if default is not available
Some distributions ship themes in separate packages so it's not sure
whether the default greeter theme ("maui") will be available.

Instead of forcing downstreams to configure sddm.conf with their theme,
we are clever enough to use the first theme available in alphabetical
order.

[ChangeLog][Daemon][Display] Load the first greeter theme in
alphabetical if the default one is not available.

Issue: #239