Fix stderr redirection in helper.cpp
The current code does not actually redirect stderr, by not redirecting
the output builds up in sddm-helper and sddm-helper grows in size

setStandardErrorFile cannot be called inside setupChildProcess as this
is after we have forked.

Instead we call code in the child process to swap the stderr fd with an
output file.

The file descriptor for stdout is swapped with a fd to /dev/null