Make PasswordBox follow user set height.
The widgets in PasswordBox have a fixed height in them. This means
that if some theme applies a custom height to PasswordBox, it is
ignored, and the default 30px is used.

This commit makes it so the widgets inside PasswordBox inherit the
container's size, so when a custom height is applied, it is followed.