clarify syntax of HideUsers & HideShells in man
Explain that the lists of users/shells should be comma-separated. When I was editing my sddm.conf, I found this quite unobvious, especially since it appeared directly under the colon-separated DefaultPath. I'd also suggest changing the default sddm.conf to reflect this, but I think that is either generated in part from this same file, or added by the Fedora packager, in my case. I'll check that out.