Switch ZSH to sh emulation mode after sourcing zlogin/zprofile
The Xsession helper script used to switch ZSHs sh emultaion mode before
sourcing the zlogin and zprofile. This caused the imports to fail if
either of the files contained ZSH specific syntax. Which in turn lead to
the session failing to start as the environment would not be setup
correctly.

This was observed with on Archlinux with SDDM 0.11.0 and KDE4/Plasma5.
ZSH was the set as login shell and Prezto ZSH configuration was
installed. In this setup the startkde script failed; the PATH was not
setup due to the sourcing failure, and the startkde script couldn't find
the utilities it needed.

Signed-off-by: Kaushal M <kshlmster@gmail.com>