Read session desktop entries with Session and pass it upon login
The greeter knows where the session entry is located (either under the
X11 or Wayland sessions path) and so it must pass this information
to the daemon.

Add a class to represent a session in order to share this information
between daemon and greater allowing to know the exact location of
the session desktop entry.

Since a relative file name can be specified for autologin session,
we need to search the desktop entry and create a Session object
with the right type.

In the long run we might want to deal with absolute desktop entry
paths and do not allow relative paths or names without extention.

Issue: #419