Support RTL languages
Translator will need to translate the QT_LAYOUT_DIRECTION string
and we will need to run the following command to update the
translation files:

  lupdate-qt5 ../../components/common/TextConstants.qml ../../src/greeter -ts <lang>.ts

SDDM now relies on that string to contain "RTL" (without quotes)
in order to set the right-to-left layout direction.

All the themes were also updated to enable layout mirroring for
right-to-left languages.

Third party themes will need to do the same to honor the RTL
setting.

Closes #329