Always create a window for each screen
Clone mode is something that should be dealt with xrandr from a
custom Xsetup script, depending on user configuration.

What we want here is to always create a window for each screen
otherwise we might end up having the greeter only on the primary screen.

Also, set window geometry to the whole screen geometry like we do
for the screen model.