A screen model for each window
we used to have only one window as big as the virtual desktop,
QML took care of creating an item for each screen by iterating on
the screen model. However we now have a better approach: we create
a view for each screen that compose the virtual desktop and thus
the QML code for each screen is responsible for drawing only its
screen.

By doing so we actually make the screen model useless, but we want
to keep it for compatibility reasons, we do however create
one for each view and expose only the screen that the view belongs to
in order to avoid creating items with different sizes.