Introduce primaryScreen property
This property holds whether the current screen is primary or not.
A QML theme can use this boolean property to hide the password prompt on
non-primary screens.

Unfortunately this property will never be changed during the greeter
life time.

With Qt 5.6 we might be able to notify when it changes.

[ChangeLog][Greeter] Expose boolean property primaryScreen to QML so that
themes can detect whether their window is on a primary screen or not.