themes: Adapt to latest screen related changes
Since screenModel no longer has useful information and
it's here just for compatibility reasons the background
now attaches to the parent which is already properly
resized by the greeter program.

Ideally we would use the primaryScreen property, but it
causes problems with mirror mode as the only view we
see is not on the primary screen.

So for now we display the login screen on every screen.