refactored ConfigReader::(is|matches)Default()
Previously, ConfigReader::isDefault didn't return if the value is
the default value but if it matches the default value. So I renamed the
method to matchesDefault() and created a new isDefault() method
that really returns if the value comes from the default value or
it was read from the config file or set.

So if the user sets a variable in the config file with the same value
as the default value, then matchesDefault() returns true, but
isDefault() returns false