Defines a users threshold to disable avatars
Defines a new config setting DisableAvatarsThreshold.
If the number of users in the system is greater than
DisableAvatarsThreshold then avatars are disabled unless
explicitly enabled with EnableAvatars. Also, UserModel
adds a new disableAvatarsThreshold property so themes can
decide if using that config setting to do further optimizations
if there's a large number of users (like for example,
not show a user list, but username/password editlines).