greeter: Set themes before running the greeter
Qt platform theme needs to be set before the QGuiApplication
is created and it can be done either with an argument or with
an environment variable.  We read the theme configuration before
running the greeter process and pass the command line argument
so we can see it from the process list.

The cursor theme can be set later but since we have a default
value from the daemon configuration and we need to set the
platform theme we might as well pass the former here.

To avoid recreating ThemeConfig and ThemeMetadata every time
the greeter theme is changed we have a setter method.