Move EnableHiDPI setting to X11 and Wayland sections
Let's have a different setting depending on the display server.

On Wayland we might have a chance to circumvent bad physical size reported
by some screen while on X11 it's harder because we cannot do that from
QPA, hence a user might want to disable HiDPI on X11 and enable it on
Wayland.