Re-use existing sessions
Most desktop environments don't support logging in as the same user
twice, especially with system changes that means the DBUS session bus is
the same.

With this patch if a user tries to log in as a user that is already
logged in we unlock and activate that session rather than creating a new
session.

This behaviour is configurable with a config option. The default matches
the current behaviour.