Add Namespaces setting
This option holds a comma-separated list of paths bound to Linux namespaces to
enter with setns() before starting the user session.

The main intention is to let user sessions start in a protected network
namespace, while the display server remains in the initial network namespace to
talk to udev, and udev remains in the initial network namespace to talk to
systemd.

[ChangeLog][General] Add "Namespaces" setting under "General" to enter specified
Linux namespaces before starting the user session