UserModel: optimize filtering out duplicate users
Instead of using naive approach of trying to detect duplicates when
inserting a new user into the list of users, which is O(n^2) complexity,
let's filter them out after the fact, which is much more efficient on
real-world data.

With this change time to load user list in our environment (over 500000
account entries) is reduced from over 1 hour to about 2 seconds.