Remove suffix for Wayland session
Some desktops like GNOME specify which windowing system is in use
with the Name entry of their desktop file.

Plasma is also headed in that direction.

For Wayland-only desktops such as Liri this information is
redundant and so is for X11-only window managers.

Do not append the Wayland suffix and let desktops handle it
themeselves.

However if two sessions has the same display name we deduplicate
appending an indicator to the Wayland counterpart, this helps
Plasma during the transition to removing the session type from
their display name.

[ChangeLog][Greeter] Remove suffix for Wayland sessions and
deduplicate display names.