UserModel: lazily load user list
In large environments loading user list can take some time and the list
is not usable anyway. The theme may opt not to show list at all, however
we create the model unconditionally. Let's lazily load data into the
model so that users disabling it do not have to pay the price.