Change location of sddm.conf.d to DATAROOTDIR
/usr/lib is not a location that should be hardcoded - depending on
distribution and architecture this can be different. So far, SDDM
does not use this path unless the above conditions happen to make
it coincide with the install location of its Qml modules.

DATAROOTDIR is defined as: Read-only architecture-independent data.
This seems to be a better fit for a system SDDM default config dir.