Fix Travis CI
We used arch-travis to build using Arch Linux in a Docker container,
but looks like it doesn't pass environment variables to the
container and this breaks the build.

Meanwhile Travis adopted Ubuntu xenial, so we can use that which
is also faster and doesn't need sudo.

Install Qt 5.8 because it's the minimum requirement, but in the
future we will be able to update to a more recent version.