Avoid adjusting active auth sessions
Display owns one Auth object. Theme's can request an auth start through
display at any time.

QProcess does not allow start to be called multiple times, so a second
m_auth->start() no-ops but by this point we've already set multiple
things. Any signals from the auth (or display's m_reuseSesion ID) are
going to be in a mixed state. When auth completes or fails we can't be
sure what we were actually authorising.

Any exploits with this appear to be hypotheticals, but I still want to
get this cleaned up with a small patch in stable.