Use avatars in FacesDir first and if not found search other locations
If Avatars are enabled and home is encrypted it takes 60s to popup the greeter at boot. The same is if copying $USER.face.icon in FacesDir since avatar is searched first in home.

The patch changes this logic. Now it search first in FacesDir and if found the greeter is shown, otherwise queries other default locations.