Cleanup sessions on exit
On exit we currently don't do any cleanup. This leaves stray processes
running. Some aren't even the responsibility of the session, such as
anything spawned in the Xsessions.d directory.

In the event of a hard reset (systemctl --restart sddm) we currently
don't even send a SIGTERM to the session leader.

By registering as a subreaper we can have the kernel automatically clean
up anything when sddm-helper is quit gracefully or unexpectedly.