Don't abort on failure to start the display server
Currently sddm logs a fatal error if X fails to start, which triggers an
abort and thus a very unclean shutdown and even a core dump.
In the case of a broken X setup, this leads to frequent service restarts
triggered by systemd which don't help and just result in core dumps and
if this fails in the handling of new seats or switchToGreeter, it'll even
kill all other sessions, potentially leading to data loss.

Logging it as critical error is enough, at least it doesn't make it worse.