FreeBSD: only set *_DESKTOP if non-empty
Only set XDG_CURRENT_DESKTOP and XDG_SESSION_DESKTOP if the xsession .desktop
file defines DesktopNames. FreeBSD xinitrc.desktop does not set it because that
may confuse the desktop session started from ~/.xinitrc.  These environment
variables are used for OnlyShowIn and NotShowIn in .desktop files.

Perhaps these two env.insert() calls can be collected in one if() but that
does change the order in which the variables are inserted, so it has been
left that way.