Fix user session logs file saving
Logs were set up in the former setupChildProcess doing manual management
of file descriptors. This was needed as we wanted to create the error
file as a regular user.

In the refactor this redirection broke, but now that sddm-helper was
changed to drop priveledges earlier, we can now use QProcess methods
directly.