Remove title bar under Weston (and possibly others)
I couldn't run sddm using Weston's fullscreen-shell.so (black screen, is it only me or is it a known bug?) Changing the compositor command line to just "weston" works, but it draws a title bar, and since Weston only supports client-side decorations it's apparently sddm's task to do this. Not sure if it breaks something under X11 (seems like it shouldn't). I'm new to sddm, let alone to C++. It's working for me, so I'm making this PR in case this it's actually the way to go.