Fix session logs
Mixing processChannelMode with setting error files does not work
reliably and our logs regressed again.

This patch only forwards when we are running the greeter.
It also groups the code together so it's more readable and hopefully
will stay working.

setStandardOutput file to nullptr was added as docs imply this is more
efficient.