Use KillMode=mixed
If xorg stops, sddm will try to create a new display, however this can
create issues at shutdown where systemd tries to kill all processes in
the cgroup.

This change changes the kill mode of sddm to mixed. With that mode, sddm
will be able to terminate the display itself. If sddm fails to do so
within reasonable interval, systemd will start killing processes in the
cgroup.