Delay restarting Display when sddm-helper doesn't terminate successfully
When the computer shuts down, sddm can try to restart Xorg and hang when
reading the display from Xorg.

This change intends to mitigate that issue by making sddm delaying
stopping display if the sddm-helper's exit code is not 0.

If the exit code is 0, the user has most likely logged out, and therefore
it should be relatively safe to restart display. Otherwise wait a second
before attempting to restart display.