Fix user switching
Make sure we don't start a new session if we are switching.
Update the terminalId (i.e. tty) Display property as we switch from the
greeter to the actual session, as they might be on different terminals now.

If we are reusing a session, also remember to switch to that session's tty.
This way we do not switch to a random open session. It's common to have 2
at most but we better fix it properly than hope our users have a predictable
workflow.