ThemeConfig: Provide typed accessors to values
Some methods are added to interpret theme.conf settings as bool, int, real and
string explicitly. Previously, most basic types were returned as string and
needed manual handling in QML code.

To allow exporting those methods to QML, it was necessary to port the class
from QVariantMap over to QQmlPropertyMap. A single method had to be added
to make it API compatible with the rest of the code.

Small breaking change: This returns more members when used with Object.keys,
e.g. the added methods and QObject's objectName.

Fixes #1097