Write logging messages to stderr if connected to a tty
standardLogger was missing a check like in journaldLogger, which caused that
it logged to the file in all cases.

Also log to stderr, that's what it's meant for. Check for a tty on stderr
instead of stdin to match that, to also make it work in processes without
forwarded stdin.