Do not force QT_WAYLAND_DISABLE_WINDOWDECORATION=1 on wayland greeters
This would put Qt in a weird state (see the issue that we are fixing)
where it would use the wrong GL configuration.

Arguably it's an error of Qt for having a bogus default pixel format but
this change should fix it but also make sure it behaves like other Qt
applications and thus more predictably.

Fixes #1441