Process all available auth messages in a loop
When QLocalSocket::readyRead is emitted, it's important to read all available
data. Otherwise the event loop is reentered while data is still available and
readyRead won't be emitted again.

This can happen in practice if sddm-helper sends a message which does not
need a reply (e.g. INFO) followed by another message before the daemon
processed the first one. In that case the daemon would never process the second
one, potentially leaving sddm-helper waiting for a reply.