Bring the greeter on primary screen to the front
When there are multiple monitors connected, sddm-greeter creates new views for each screen, and activates the primary one to focus on its password field. However, if the monitors are initially setup to display the same content, different views will overlap with each other. The views created by non-primary screens may be placed on top of the primary view. In such cases, the primary view will be focused but not visible to the user.

This is fixed by calling `view->raise` for the primary screen. The primary view should be raised whenever a new screen is added, in order to avoid being covered by the newly created view.