Allow building sddm-greeter for both Qt 5 and Qt 6 in one go
When building with Qt 6, greeter is by default built and installed for both
Qt 5 and Qt 6. When building with Qt 5, the greeter can optionally be built
and installed for Qt 6 too.

This means QT_IMPORTS_DIR/QML_INSTALL_DIR and COMPONENTS_TRANSLATION_DIR are
now Qt version specific and can't be overridden anymore.