Drop unnecessary ECM dependency and dead uninstall target
ECM was added by 6ef91b0b because of some potential dependency needed for
Wayland, which was not added until now.

The commit also changed the "uninstall" target to not be created if
ECM >= 1.7.0. ECM itself doesn't register an uninstall target automatically
either, which means this is dead code. Just remove it.