Move autologin fallback (handle auth error) from `slotAuthError` to `slotAuthenticationFinished`
There may be a case when pam authentication error not only of the `auth` type, but also, for example, of the `session` type fall into the `slotAuthError` during autologin. For example, a pam module error in the file `/etc/pam.d/postlogin` already during the loaded session. As a result, a fallback (`handleAutologinFailure`) is triggered that loads greeter (sddm-theme) already on top of the loaded session, which leads to unpleasant consequences.

In my opinion, this is incorrect, and I think it's worth transferring the fallback to the `slotAuthenticationFinished` method and calling fallback if `success` = false, then the fallback will trigger exactly when an error of the `auth` type occurred, i.e. when the session really wasn't loaded and the greeter should be loaded to avoid black screen with only cursor.