Correctly integrate the greeter with layer-shell
When using Wayland as the display server it is possible and very
likely that we are requested to use the layer-shell protocol to
display the greeter, as it better fits its windowing model.
This is selected with QT_WAYLAND_SHELL_INTEGRATION=layer-shell.

With Plasma 6 LayerShellQt has changed the default keyboard policy
to not receive any inputs, now following the protocol spec.
That breaks the ability type in a username/password, unless we
request it to enable keyboard events again.