fix(auth): do not cast exit status if `QProcess` is not exited normally
A QProcess::CrashExit happen in most cases when a signal killed it. And
returning from a signal-killed process on Linux makes QProcess:finished
return the signal, which is a really bad idea to cast into a random
enum.

Qt also mentions that `exitCode` is only valid for QProcess::NormalExit
invocations

Signed-off-by: tsrk. <tsrk@tsrk.me>